Move integration test runner code to devtools_shared
