+++ packages/devtools_shared/lib/devtools_test_utils.dart
@@ -3,5 +3,8 @@
 // found in the LICENSE file.
 
 export 'src/test/chrome.dart';
+export 'src/test/chrome_driver.dart';
 export 'src/test/cli_test_driver.dart';
+export 'src/test/integration_test_runner.dart';
+export 'src/test/io_utils.dart';
 export 'src/test/test_utils.dart';
