+++ packages/devtools_shared/pubspec.yaml
@@ -1,7 +1,7 @@
 name: devtools_shared
 description: Package of shared Dart structures between devtools_app, dds, and other tools.
 
-version: 4.0.1
+version: 4.1.0
 
 repository: https://github.com/flutter/devtools/tree/master/packages/devtools_shared
